build (multer): gestion routes post, put, delete

+++ backend/controllers/ctrlBooks.js
@@ -1,4 +1,5 @@
 const Books = require("../models/modelBooks");
+const fs = require('fs');
 
 exports.getAllBooks = (req, res, next) => {
     Books.find()
@@ -7,7 +8,7 @@ exports.getAllBooks = (req, res, next) => {
   };
 
   exports.createBooks = (req, res, next) => {
-    const booksObject = JSON.parse(req.body.books);
+    const booksObject = JSON.parse(req.body.book);
     delete booksObject._id;
     delete booksObject._userId;
     const books = new Books({
@@ -27,13 +28,42 @@ exports.getAllBooks = (req, res, next) => {
   };
 
   exports.modifyBooks = (req, res, next) => {
-    Books.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-      .then(() => res.status(200).json({ message: 'Livre modifié !'}))
-      .catch(error => res.status(400).json({ error }));
-  };
+      const booksObject = req.file ? {
+          ...JSON.parse(req.body.book),
+          imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+      } : { ...req.body };
+    
+      delete booksObject._userId;
+      Books.findOne({_id: req.params.id})
+          .then((book) => {
+              if (book.userId != req.auth.userId) {
+                  res.status(401).json({ message : 'Not authorized'});
+              } else {
+                  Books.updateOne({ _id: req.params.id}, { ...booksObject, _id: req.params.id})
+                  .then(() => res.status(200).json({message : 'Livre modifié!'}))
+                  .catch(error => res.status(401).json({ error }));
+              }
+          })
+          .catch((error) => {
+              res.status(400).json({ error });
+          });
+   };
 
   exports.deleteBooks = (req, res, next) => {
-    Books.deleteOne({ _id: req.params.id })
-      .then(() => res.status(200).json({ message: 'Livre supprimé !'}))
-      .catch(error => res.status(400).json({ error }));
+      Books.findOne({ _id: req.params.id})
+          .then(book => {
+              if (book.userId != req.auth.userId) {
+                  res.status(401).json({message: 'Not authorized'});
+              } else {
+                  const filename = book.imageUrl.split('/images/')[1];
+                  fs.unlink(`images/${filename}`, () => {
+                      Books.deleteOne({_id: req.params.id})
+                          .then(() => { res.status(200).json({message: 'Livre supprimé !'})})
+                          .catch(error => res.status(401).json({ error }));
+                  });
+              }
+          })
+          .catch( error => {
+              res.status(500).json({ error });
+          });
   };
